refactor: Use makeTheme to create MuiCozyThemes
